+++ types/CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `OutputResponse` enum with `Json` and `Raw` variants;
 - `AliasOutput::alias_id_non_null` and `NftOutput::nft_id_non_null` methods;
+- serde aliases to `ProtocolParameters` fields for backwards compatibility;
 
 ### Changed