+++ types/CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AliasAddress::into_alias_id` method;
 - `Output::{is_treasury, is_basic, is_alias, is_foundry, is_nft}` methods;
 - `RentStructure::new` and getters;
+- serde aliases to `ProtocolParameters` fields for backwards compatibility;
 
 ### Changed
 
@@ -42,7 +43,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `OutputResponse` enum with `Json` and `Raw` variants;
 - `AliasOutput::alias_id_non_null` and `NftOutput::nft_id_non_null` methods;
-- serde aliases to `ProtocolParameters` fields for backwards compatibility;
 
 ### Changed